File to load the data from CSV files
This is the file to load the data from the CSV files. The CSV files are in a folder called data and I will upload that to the GitHub as well. Each function loads data from each file corresponding to the name of the function in a list of lists where each inner list is a row of the csv file. The last function is called load_all which loads all csv files at once and returns them in a dictionary of key strings and values of the data in the list of lists format. So to load all data at once and use specific data you can index the dictionary like:  load_all()['wages'] which will return the wages csv data in a list of lists.